Stop PPCoin addresses overflowing the panel
This was a problem on the Project Sponsors panel as demonstrated at http://peer4commit.com/projects/1. This change makes the entire block gain a scroll bar if necessary, so that the entire value is still accessible, but it is not in the way, overflowing onto the top of the main content.

This was mainly an issue for a viewport width in the range 992–1199px, though with a significant percentage of wider characters (e.g. lots of W's, not many i's) it could still appear outside that range.